refactor: move ecstore rebalance entry flow
